requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community

# OpenAI Integration
langchain-openai
openai

# Groq Integration
langchain_groq

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
sentence-transformers 
torch
pydantic
pypdf
unstructured

#UI
streamlit
grandalf